+++ CHANGELOG.md
@@ -18,8 +18,9 @@ Group changes to describe their impact on the project, as follows:
 ### [Added]
 - **Proxy:**
     - Push notification delivery can now be delegated to a FlexiAPI server instance.
-- **Conference server:** add the 'no-rtp-timeout' parameter that allows to set the delay before the call is automatically
-  hung up because no RTP data is received.
+- **Conference server:**
+    - Add the 'no-rtp-timeout' parameter that allows to set the delay before the call is automatically hung up because no RTP data is received.
+    - Add support for end-to-end conference encryption by installing the EKT server plugin. For customers under a proprietary license, this functionality is under a specific license.
 
 ### [Changed]
 - **Pusher:**